two tasks

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int i, n;
+	long int i, n;
 
 	i = 2;
 	n = 612852475143;
